+++ code_test/src/backjoon/done/No1085.java
@@ -1,4 +1,4 @@
-package backjoon;
+package backjoon.done;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
