+++ code_test/src/backjoon/done/No11653_2.java
@@ -1,16 +1,15 @@
-package backjoon;
+package backjoon.done;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
 
-public class No11653 {
+public class No11653_2 {
     public static void main(String[] args) throws Exception {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         int N = Integer.parseInt(br.readLine());
-        List<Integer> list = new ArrayList<>();
 
         if (N == 1) {
             return;
@@ -18,17 +17,11 @@ public class No11653 {
 
         for (int i = 2; i <= N; i++) {
             while (N % i == 0) {
-                list.add(i);
+                System.out.println(i);
                 N = N / i;
             }
         }
 
-       if (list.isEmpty()) {
-            System.out.println(N);
-        } else
-            for (int i : list) {
-                System.out.println(i);
-            }
 
     }
 }
