+++ code_test/src/backjoon/done/No24262.java
@@ -1,0 +1,9 @@
+package backjoon.done;
+
+public class No24262 {
+    public static void main(String[] args) throws Exception {
+        // 단일 연산의 경우 연산을 하는 횟수만큼 시간 복잡도가 고정적으 나온다.
+        System.out.println(1);
+        System.out.println(0);
+    }
+}
