+++ code_test/src/backjoon/done/No3009.java
@@ -1,4 +1,4 @@
-package backjoon;
+package backjoon.done;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
