+++ code_test/src/backjoon/done/No9063.java
@@ -1,4 +1,4 @@
-package backjoon;
+package backjoon.done;
 
 import java.util.Arrays;
 import java.util.Scanner;
