+++ code_test/src/backjoon/done/No9506_2.java
@@ -1,4 +1,4 @@
-package backjoon;
+package backjoon.done;
 
 import java.util.Scanner;
