+++ code_test/src/backjoon/july5/No24263.java
@@ -1,17 +1,13 @@
-package backjoon;
+package backjoon.july5;
 
 import java.io.BufferedReader;
-import java.io.IOException;
 import java.io.InputStreamReader;
 
-public class No27323 {
+public class No24263 {
     public static void main(String[] args) throws Exception {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
-        int a = Integer.parseInt(br.readLine());
-        int b = Integer.parseInt(br.readLine());
-
-        System.out.println(a*b);
-
+        System.out.println(Integer.parseInt(br.readLine()));
+        System.out.println(1);
     }
 }
