+++ code_test/src/backjoon/july5/No24264.java
@@ -1,0 +1,14 @@
+package backjoon.july5;
+
+import java.util.Scanner;
+
+public class No24264 {
+    public static void main(String[] args){
+        Scanner sc = new Scanner(System.in);
+
+        long a = sc.nextLong();
+
+        System.out.println(a*a);
+        System.out.println(2);
+    }
+}
