+++ code_test/src/backjoon/july5/No24265.java
@@ -1,7 +1,7 @@
-package backjoon;
+package backjoon.july5;
 
-public class No24262 {
+public class No24265 {
     public static void main(String[] args) throws Exception {
-        
+
     }
 }
