+++ code_test/src/backjoon/july5/No27323.java
@@ -1,0 +1,17 @@
+package backjoon.july5;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+
+public class No27323 {
+    public static void main(String[] args) throws Exception {
+        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+
+        int a = Integer.parseInt(br.readLine());
+        int b = Integer.parseInt(br.readLine());
+
+        System.out.println(a*b);
+
+    }
+}
